Introduced integer_math.hpp.
Moved some of the generic functions from the implementation of
binary_page_memory_resource into this new place. While adding
some additional functions, which would be needed for
pool_memory_resource.
